Update internal token endpoint (#182)
* Quick changes to get prod working

* Reuse more code

* Delete old files

* Remove more code

* Fix tests

* Fix bug

* Improve API error handling some more

* Update changelog [skip ci]

* Poke CI
