--- VideoApp/VideoApp/API/Core/APIErrorResponse.swift
@@ -1,30 +1,0 @@
-//
-//  Copyright (C) 2020 Twilio, Inc.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-struct APIErrorResponse: Decodable {
-    struct Error: Decodable {
-        enum Message: String, Decodable {
-            case passcodeExpired = "passcode expired"
-            case passcodeIncorrect = "passcode incorrect"
-        }
-        
-        let message: Message
-    }
-    
-    let error: Error
-}
