+++ VideoApp/VideoApp/Stores/AppSettings/AppSettingsStore.swift
@@ -19,7 +19,6 @@ import Foundation
 protocol AppSettingsStoreWriting: LaunchStore {
     var environment: Environment { get set }
     var videoCodec: VideoCodec { get set }
-    var topology: Topology { get set }
     var userIdentity: String { get set }
     var isTURNMediaRelayOn: Bool { get set }
     var areInsightsEnabled: Bool { get set }
@@ -31,7 +30,7 @@ protocol AppSettingsStoreWriting: LaunchStore {
     var maxSubscriptionBitrate: Int? { get set }
     var dominantSpeakerPriority: TrackPriority { get set }
     var trackSwitchOffMode: TrackSwitchOffMode { get set }
-    var remoteRoomType: CommunityCreateTwilioAccessTokenResponse.RoomType? { get set }
+    var remoteRoomType: CreateTwilioAccessTokenResponse.RoomType? { get set }
     var clientTrackSwitchOffControl: ClientTrackSwitchOffControl { get set }
     var videoContentPreferencesMode: VideoContentPreferencesMode { get set }
     func reset()
@@ -40,7 +39,6 @@ protocol AppSettingsStoreWriting: LaunchStore {
 class AppSettingsStore: AppSettingsStoreWriting {
     @Storage(key: makeKey("environment"), defaultValue: Environment.production) var environment: Environment
     @Storage(key: makeKey("videoCodec"), defaultValue: defaultVideoCodec) var videoCodec: VideoCodec
-    @Storage(key: makeKey("topology"), defaultValue: Topology.group) var topology: Topology
     @Storage(key: makeKey("userIdentity"), defaultValue: "") var userIdentity: String
     @Storage(key: makeKey("isTURNMediaRelayOn"), defaultValue: false) var isTURNMediaRelayOn: Bool
     @Storage(key: makeKey("areInsightsEnabled"), defaultValue: true) var areInsightsEnabled: Bool
@@ -52,7 +50,7 @@ class AppSettingsStore: AppSettingsStoreWriting {
     @Storage(key: makeKey("maxSubscriptionBitrate"), defaultValue: nil) var maxSubscriptionBitrate: Int?
     @Storage(key: makeKey("dominantSpeakerPriority"), defaultValue: TrackPriority.serverDefault) var dominantSpeakerPriority: TrackPriority
     @Storage(key: makeKey("trackSwitchOffMode"), defaultValue: TrackSwitchOffMode.serverDefault) var trackSwitchOffMode: TrackSwitchOffMode
-    @Storage(key: makeKey("remoteRoomType"), defaultValue: nil) var remoteRoomType: CommunityCreateTwilioAccessTokenResponse.RoomType?
+    @Storage(key: makeKey("remoteRoomType"), defaultValue: nil) var remoteRoomType: CreateTwilioAccessTokenResponse.RoomType?
     @Storage(key: makeKey("clientTrackSwitchOffControl"), defaultValue: .sdkDefault) var clientTrackSwitchOffControl: ClientTrackSwitchOffControl
     @Storage(key: makeKey("videoContentPreferencesMode"), defaultValue: .sdkDefault) var videoContentPreferencesMode: VideoContentPreferencesMode
