+++ VideoApp/VideoApp/Stores/Auth/AuthError.swift
@@ -28,11 +28,7 @@ enum AuthError: Error, Equatable {
 
     init(apiError: APIError) {
         switch apiError {
-        case .decodeError: self = .unknown
         case let .message(message): self = .message(message: message)
-        case .notConnectedToInternet: self = .networkError
-        case .passcodeExpired: self = .passcodeExpired
-        case .passcodeIncorrect: self = .passcodeIncorrect
         }
     }
 }
