+++ VideoApp/VideoApp/Stores/TwilioAccessToken/TwilioAccessTokenStore.swift
@@ -19,3 +19,45 @@ import Foundation
 protocol TwilioAccessTokenStoreReading: AnyObject {
     func fetchTwilioAccessToken(roomName: String, completion: @escaping (Result<String, APIError>) -> Void)
 }
+
+class TwilioAccessTokenStore: TwilioAccessTokenStoreReading {
+    private let api: APIRequesting
+    private let appSettingsStore: AppSettingsStoreWriting
+    private let authStore: AuthStoreWriting
+    private let remoteConfigStore: RemoteConfigStoreWriting
+
+    init(
+        api: APIRequesting,
+        appSettingsStore: AppSettingsStoreWriting,
+        authStore: AuthStoreWriting,
+        remoteConfigStore: RemoteConfigStoreWriting
+    ) {
+        self.api = api
+        self.appSettingsStore = appSettingsStore
+        self.authStore = authStore
+        self.remoteConfigStore = remoteConfigStore
+    }
+    
+    func fetchTwilioAccessToken(roomName: String, completion: @escaping (Result<String, APIError>) -> Void) {
+        authStore.refreshIDToken { [weak self] in
+            guard let self = self else { return }
+
+            let request = CreateTwilioAccessTokenRequest(
+                passcode: self.authStore.passcode ?? "",
+                userIdentity: self.appSettingsStore.userIdentity.nilIfEmpty ?? self.authStore.userDisplayName,
+                createRoom: true,
+                roomName: roomName
+            )
+            
+            self.api.request(request) { [weak self] result in
+                guard let self = self else { return }
+                
+                if let roomType = try? result.get().roomType {
+                    self.remoteConfigStore.roomType = roomType
+                }
+                
+                completion(result.map { $0.token })
+            }
+        }
+    }
+}
